Changing enviorment to allow for less abstraction with regards to the qaRounter:
