+++ gainsboro-atelier-main/imageAPI/imageAPI.js
@@ -1,24 +1,24 @@
-const cloudinary = require("cloudinary");
-const {cloudinaryInfo} = require("../config.js");
+// const cloudinary = require("cloudinary");
+// const {cloudinaryInfo} = require("../config.js");
 
-const {cloud_name, api_key, api_secret} = cloudinaryInfo;
+// const {cloud_name, api_key, api_secret} = cloudinaryInfo;
 
-cloudinary.config({
-  cloud_name,
-  api_key,
-  api_secret,
-})
+// cloudinary.config({
+//   cloud_name,
+//   api_key,
+//   api_secret,
+// })
 
-const uploadImages = (files) => {
-  const urlArr = files.map(file => {
-    return cloudinary.uploader.upload(file.path)
-    .then(data => {
-      var url = data.secure_url;
-      return url;
-    })
-    .catch(err => {throw err})
-  })
-  return Promise.all(urlArr)
-}
+// const uploadImages = (files) => {
+//   const urlArr = files.map(file => {
+//     return cloudinary.uploader.upload(file.path)
+//     .then(data => {
+//       var url = data.secure_url;
+//       return url;
+//     })
+//     .catch(err => {throw err})
+//   })
+//   return Promise.all(urlArr)
+// }
 
-module.exports = uploadImages;
+// module.exports = uploadImages;
