+++ gainsboro-atelier-main/server/server.js
@@ -31,7 +31,7 @@ var connection = mysql.createConnection({
 
 connection.connect(function(err) {
     if (err) throw err;
-    console.log("Connected to the Gosh Darn DataBase!");
+    // console.log("Connected to the Gosh Darn DataBase!");
 });
 
 
@@ -57,7 +57,6 @@ app.get('/overview/:product_id', (req, res) => {
 
 app.post('/cart', (req, res)=>{
   const {sku_id} = req.body;
-  console.log('lksjd');
 
   const url = `${apiHost}/cart`;
   const data = {
@@ -200,7 +199,60 @@ app.put('/reviews/:review_id/report', (req, res) => {
   .catch(err => res.sendStatus(500))
 })
 
-app.use('/qa',  qaRouter);
+// app.use('/qa',  qaRouter);
+
+var getAnswerQuestionData = function(answersQuery, data, index, dataLength, callback) {
+  connection.query(answersQuery, function (error, answerData) {
+      if (error) res.sendStats(500);
+      data[index]['AnswerData'] = answerData;
+      if (dataLength === index) {
+          callback(null, data);
+      }
+  })
+}
+
+app.route('/questions')
+  .get((req, res) => {
+      var {product_id} = req.query;
+      if (product_id === null || product_id === undefined) product_Id = 71967;
+      var url = `${apiHost}/qa/questions?product_id=${product_id}`;
+      var queryStatement = `Select * from Questions where Product_Id = ${product_id}`;
+      connection.query(queryStatement, function (error, data){
+          if (error) res.sendStatus(500);
+          var dataLength = data.length - 1;
+          const loop = async () => {
+              var count = 0;
+              for (const index of data) {
+                  var question_Id = data[count]['Question_Id'];
+                  var answersQuery = `Select * from Answers where Question_Id = ${question_Id}`
+                  await getAnswerQuestionData(answersQuery, data, count, dataLength, function (error, data) {
+                      res.sendStatus(200).send(data);
+                  });
+                  count++;
+              }
+          }
+          loop()
+      })
+  })
+  .post(jsonParser, (req, res) => {
+    const {body, name, email, product_id} = req.body;
+    var date =  Date.now();
+    var questionInsert = `SELECT MAX (Question_Id) from Questions`;
+    connection.query(questionInsert, function (error, data){
+        if (error) res.sendStatus(500);
+        const data2 = data['MAX(Question_Id)'];
+        var questionMax = data[0]['MAX (Question_Id)'] + 1;
+        var sqlInsertCode = `INSERT INTO questions (Question_Id, Question_Body, Question_Date, Asker_Name, Product_Id, Reported, Helpful) VALUES (${questionMax}, "${body}", "${date}", "${name}", ${product_id}, ${0}, ${0})`;
+        connection.query(sqlInsertCode, function(err, data) {
+            if (err) res.sendStatus(500);
+            res.send('data')
+        })
+    })
+})
+
+
+
+
 
 app.post('/answers/answer_id/helpful', (req, res) => {
     var answer_id = req.body.answer_Id;
@@ -215,7 +267,6 @@ app.post('/answers/answer_id/helpful', (req, res) => {
 app.post('/answers/answer_id/report', (req, res) => {
     var answer_id = req.body.answer_Id;
     var question_id = req.body.question_Id;
-    console.log(answer_id, question_id, 'hi')
     var queryStatement = `UPDATE Answers set Reported = 1 where Answer_Id = ${answer_id} AND Question_Id = ${question_id}`;
     connection.query(queryStatement, function (error, data){
       if (error) {
@@ -234,20 +285,15 @@ app.post('/questions/question_id/helpful', (req, res) => {
       if (error) {
         res.sendStatus(500);
       } else {
-        console.log('Successful Questions Id Update')
         res.send(data);
       }
   })
 })
 
 
-
-
 app.post('/interactions', jsonParser, (req, res) => {
   var body = req.body;
-  console.log(body);
   var url = `${apiHost}/interactions`;
-  console.log(url)
   axios.post(url, body, {
       'content-type': 'application/json',
       headers: {
@@ -255,10 +301,9 @@ app.post('/interactions', jsonParser, (req, res) => {
       }
   })
   .then(data => {
-      res.send(data.data)
+      res.sendStatus(200).send(data.data)
   })
   .catch(err => {
-      console.log('Error in the interactions')
       res.sendStatus(500)
   })
 })
@@ -268,5 +313,7 @@ app.get('*', function (request, response) {
 });
 
 app.listen(PORT, () => {
-  console.log(`connected to port ${PORT}`);
+  // console.log(`connected to port ${PORT}`);
 });
+
+module.exports = app;
